pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "python-multipart>=0.0.9",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "rank-bm25>=0.2.2",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.0",
    "beautifulsoup4>=4.12.0",
]
